moved field constants to its own file
